Move tests and remove old module
